+++ scripts/iblone/purge_rig_data.py
